fix(houdini-react): resolve circular dependencies
houdini-react/src/plugin/index.ts -> houdini-react/src/plugin/vite.tsx
houdini-react/src/runtime/routing/Router.tsx -> houdini-react/src/runtime/hooks/useDocumentHandle.ts
houdini-react/src/runtime/hooks/useDocumentStore.ts -> index -> houdini-react/src/runtime/routing/Router.tsx
