+++ packages/houdini-react/src/plugin/index.ts
@@ -9,7 +9,6 @@ import {
 	type Document,
 	type Config,
 	type Plugin,
-	type ProjectManifest,
 } from 'houdini'
 import path from 'node:path'
 import { loadEnv } from 'vite'
@@ -19,8 +18,7 @@ import { format_router_manifest } from './codegen/router'
 import { extractDocuments } from './extract'
 import { transformFile } from './transform'
 import vite_plugin from './vite'
-
-export let manifest: ProjectManifest
+import { manifest, setManifest } from './state'
 
 export const hooks: Plugin = async () => ({
 	order: 'core',
@@ -32,7 +30,7 @@ export const hooks: Plugin = async () => ({
 	// we generate anything
 	async beforeGenerate({ config }) {
 		try {
-			manifest = await load_manifest({ config })
+			setManifest(await load_manifest({ config }))
 		} catch (e) {
 			console.log('something went wrong: ' + (e as Error).message)
 			return
@@ -328,6 +326,3 @@ export default plugin('houdini-react', hooks)
 
 export type HoudiniReactPluginConfig = {}
 
-export function setManifest(newManifest: ProjectManifest): void {
-	manifest = newManifest
-}
