+++ packages/houdini-react/src/runtime/hooks/useDocumentHandle.ts
@@ -1,3 +1,5 @@
+import React from 'react'
+
 import type { DocumentStore } from '$houdini/runtime/client'
 import { extractPageInfo } from '$houdini/runtime/lib/pageInfo'
 import { cursorHandlers, offsetHandlers } from '$houdini/runtime/lib/pagination'
@@ -13,9 +15,8 @@ import type {
 	DocumentArtifact,
 	QueryArtifact,
 } from '$houdini/runtime/lib/types'
-import React from 'react'
 
-import { useClient, useLocation, useSession } from '../routing/Router'
+import { useClient, useSession, useLocation } from '../routing/hooks'
 
 export function useDocumentHandle<
 	_Artifact extends QueryArtifact,
